requirements.txt
django
djangorestframework
django-widget-tweaks
django-session-timeout
djangorestframework-datatables
django-select2
dash
dash-bootstrap-components
dash-core-components
dash-html-components
dash-renderer
dash-table
django-appconf
django-plotly-dash
dpd-components
Fiona
fonttools
geojson
geopandas
itsdangerous
Jinja2
jmespath
kiwisolver
mapbox
mapboxgl
MarkupSafe
matplotlib
msgpack
munch
numpy
packaging
pandas
Pillow
plotly
polyline
pyparsing
pyproj
python-dateutil
pytz
s3transfer
shapely
six
sqlparse
tenacity
tzdata
uritemplate
urllib3
Werkzeug
redis
channels_redis
